draw: Silencium compiler warnings on Windows.

+++ src/gallium/auxiliary/draw/draw_pipe_offset.c
@@ -123,7 +123,7 @@ static void offset_first_tri( struct draw_stage *stage,
 {
    struct offset_stage *offset = offset_stage(stage);
 
-   offset->units = stage->draw->rasterizer->offset_units * stage->draw->mrd;
+   offset->units = (float) (stage->draw->rasterizer->offset_units * stage->draw->mrd);
    offset->scale = stage->draw->rasterizer->offset_scale;
 
    stage->tri = offset_tri;
